Stop sending carriage returns
The carriage return at the end of each line was being interpreted as an additional command. This resulted in each response from a grbl board containing 2 ok's for each command send. This caused a sufficiently large gcode program to eventually overflow the buffer of the grbl controller resulting in a "error: Modal group violation".