Rate limit display updates to the command log
Update the command log at most 5 times a second.

The command log internally can be updated very quickly if the connected
machine consumes the commands quickly. If we tried to update the
displayed log that fast, the machine would quickly get bogged down with
rendering and won’t send commands at the rate we need.

This change partially addresses issue #33.