Speed up command queue processing a bit
fix #33

Now that we rate limit updates to the command log UI, we can speed up
command processing without completely hosing lesser machines (I hope).

Note, there may still be issues if a machine is able to processes
commands with an average of less than 10ms (which is insanely fast for
a physical mechanism to move and likely means the commands aren’t
moving it very far).